Static analysis fix
#rb none
#rnx

#ROBOMERGE-AUTHOR: per.larsson
#ROBOMERGE-SOURCE: CL 18189262 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v895-18170469)

[CL 18189290 by per larsson in ue5-release-engine-staging branch]

+++ Engine/Source/Developer/DerivedDataCache/Private/ZenDerivedDataBackend.cpp
@@ -609,7 +609,7 @@ void FZenDerivedDataBackend::Get(
 	ForEachBatch(CacheRecordBatchSize, Keys.Num(),
 		[this, &Keys, &Context, &Policy, &Owner, &OnComplete, &TotalCompleted](int32 BatchFirst, int32 BatchLast)
 	{
-		TRACE_COUNTER_ADD(ZenDDC_Get, int64((BatchLast - BatchFirst) + 1));
+		TRACE_COUNTER_ADD(ZenDDC_Get, int64(BatchLast) - int64(BatchFirst) + 1);
 		COOK_STAT(auto Timer = UsageStats.TimeGet());
 		
 		FCbWriter BatchRequest;
@@ -757,7 +757,7 @@ void FZenDerivedDataBackend::GetChunks(
 	ForEachBatch(CacheChunksBatchSize, SortedChunks.Num(),
 		[this, &SortedChunks, &Context, &Owner, &OnComplete, &TotalCompleted](int32 BatchFirst, int32 BatchLast)
 	{
-		TRACE_COUNTER_ADD(ZenDDC_Get, int64((BatchLast - BatchFirst) + 1));
+		TRACE_COUNTER_ADD(ZenDDC_Get, int64(BatchLast) - int64(BatchFirst) + 1);
 		COOK_STAT(auto Timer = UsageStats.TimeGet());
 
 		FCbWriter BatchRequest;
